embedder/omai-zip/requirements.txt
transformers>=4.30.0
sentence-transformers>=2.2.0
huggingface_hub>=0.20.0
pymongo>=4.5.0
numpy>=1.24.0
tqdm>=4.65.0
python-dotenv>=1.0.0
